requirements.txt
Django>=5.2
djangorestframework>=3.16
django-cors-headers>=4.9
django-resized>=1.0.3
Pillow>=12.0
gunicorn>=22.0
whitenoise>=6.8
